Reference figures, tables and equations in R Markdown documents

These functions return closures that allow for constructing a series of numbered items and to reference them. The number is create the first time a label is encountered, and provided again for further use of the same label.

new_labelling(
  type = c("arabic", "roman"),
  prefix = "",
  string = "_**Figure ####:** @@@@_"
)

newLabelling(
  type = c("arabic", "roman"),
  prefix = "",
  string = "_**Figure ####:** @@@@_"
)

fig(label, text, reset = FALSE)

tab(label, text, reset = FALSE)

eq(label, text, reset = FALSE)

Arguments

type
The type of numbering (arabic or roman).

prefix
The text to use to prefix the number (nothing, by default).

string
The string prototyping the legend, with @@ being a placeholder for the text, #### as a placeholder for the number, or %%%% as a placeholder for numbered examples (used by eq()).

label
A short string uniquely identifying the item within the collection.

text
The text for the legend (can contain R Markdown formatting). In the case of eq(), this is the LaTeX code for the equation itself.

reset
If TRUE, the collection is reset. Useful if you want to restart numbering at the beginning of each chapter.

Value

The item labelling (legend, ...) if `text =`` is provided, or the number associated with the item otherwise. new_labelling() creates a new labelling function, which has the same arguments as fig().

Details

A new labelling type is created using new_labelling() which is a function factory (a function that creates functions).

Examples

# Note that those instructions are supposed to be included in R Markdown
# documents using the inline R expression syntax: `r fig("label", "<text>")`.

# Say you have a first figure labelled 'first-plot'.
# The first call creates its legend, and the second one references it
fig("first-plot", "An example of a gpplot2 scatterplot.")

#> [1] "_**Figure 1:** An example of a gpplot2 scatterplot._"

fig("first-plot")

#> [1] "1"

# A second figure labelled 'a-picture', but with reference appearing first
fig("a-picture")

#> [1] "2"

# The second call creates the legend for the picture
fig("a-picture", "This is my most beautiful picture")

#> [1] "_**Figure 2:** This is my most beautiful picture_"

# Same for a table labelled 'anova-table'
tab("anova-table", "ANOVA for my wonderful dataset.")

#> [1] "_**Table 1:** ANOVA for my wonderful dataset._"

# ... and further in the text, a reference to that table:
tab("anova-table")

#> [1] "1"